Appロジックを分離し時間制御と千日手判定の単体テストを追加
